backend/requirements.txt
flask
flask-restful
flask-cors
flask-sqlalchemy
psycopg2
gunicorn
requests
pyopenssl
